+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) convention.
 
+## [3.0.5] - 07-9-2024
+
+### Added
+
+- `--force` arg to model deletion
+- error handling for when user was not found in logs
+
+### Changed
+
+- Include `tests/` within the package to allow reuse in parent package
+
 ## [3.0.4] - 07-1-2024
 
 ### Removed
@@ -109,6 +120,7 @@ Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and
 
 - switched to `pathlib` for file path parsing
 
+[3.0.5]: https://github.com/A-Baji/discordAI-modelizer/compare/3.0.4...3.0.5
 [3.0.4]: https://github.com/A-Baji/discordAI-modelizer/compare/3.0.3...3.0.4
 [3.0.3]: https://github.com/A-Baji/discordAI-modelizer/compare/3.0.2...3.0.3
 [3.0.2]: https://github.com/A-Baji/discordAI-modelizer/compare/3.0.1...3.0.2